#81 [FIX] 클라이언트 컴포넌트 분리로 오류 해결
